+++ packages/front-end/components/Experiment/VisualChangesetTable.tsx
@@ -14,8 +14,8 @@ import { appendQueryParamsToURL } from "@/services/utils";
 import { useAuth } from "@/services/auth";
 import VisualChangesetModal from "@/components/Experiment/VisualChangesetModal";
 import EditDOMMutationsModal from "@/components/Experiment/EditDOMMutationsModal";
-import LinkedChange from "@/components/Experiment/LinkedChange";
-import LinkedChangeVariationRows from "@/components/Experiment/LinkedChangeVariationRows";
+import LinkedChange from "@/components/Experiment/LinkedChanges/LinkedChange";
+import LinkedChangeVariationRows from "@/components/Experiment/LinkedChanges/LinkedChangeVariationRows";
 import EnvironmentStatesGrid from "@/components/Experiment/LinkedChanges/EnvironmentStatesGrid";
 import Badge from "@/ui/Badge";
 import Button from "@/ui/Button";
